requirements.txt
langgraph>=0.2
langchain-openai>=0.1
langchain-core>=0.2
langchain-mcp-adapters>=0.1
openai>=1.0
mcp
fastapi
uvicorn[standard]
matplotlib
pandas
python-dotenv
docker>=7.0.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
alembic>=1.13
python-jose[cryptography]>=3.3
httpx>=0.27
psycopg2-binary>=2.9
pytest
pytest-asyncio
testcontainers[postgres]>=4.0
